tests: Rename 'aux/' directory to 'auxiliary/'.
Unfortunately, 'aux' is one of the Forbidden Names. [0]

[0] https://msdn.microsoft.com/en-us/library/windows/desktop/aa365247(v=vs.85).aspx
