+++ src/__tests__/auxiliary/lolex.js
